// src/enums/Colors.ts
export enum CyberpunkSynthWaveColors {
    CleanPoolBlue = "#59CBE8",
    Maldives = "#00BCE1",
    PinkBite = "#E93CAC",
    Violet = "#8A2BE2",
    NeonGreen = "#39FF14",
    SapphireSplendour = "#1E22AA",
    ShadowPlanet = "#201547",
    MidnightDreams = "#051C2C",
    Black = "hsl(0, 0%, 4%)"
}

export enum MatrixColors {
    White = "#FFFFFF",
    Black = "#020204",
    Green_1 = "#204829",
    Green_2 = "#22b455",
    Green_3 = "#22b455",
    Green_4 = "#92e5a1",
}

export enum TechnoColors {
    NeonGreen = "#39FF14",
    ElectricBlue = "#7DF9FF",
    HotPink = "#FF69B4",
    CyberYellow = "#FFD700",
    UltraViolet = "#6A0DAD",
    LaserRed = "#FF073A",
    DigitalOrange = "#FF4500",
    VirtualCyan = "#00FFFF",
    HologramWhite = "#F0FFFF",
    MatrixBlack = "#0B0C10",
}

export enum BladeRunnerColors {
    NeonRed = "#FF073A",
    ElectricBlue = "#7DF9FF",
    DeepPurple = "#6A0DAD",
    CyberYellow = "#FFD700",
    DarkGray = "#2E2E2E",
    NeonGreen = "#39FF14",
    HologramWhite = "#F0FFFF",
    MidnightBlack = "#0B0C10",
}

export enum  GlitchColors {
    Red = "#FF0000",
    Green = "#00FF00",
    Blue = "#0000FF"
}

export enum NoSignalColors {
    White = "#FFFFFF",
    Yellow = "#FFFF00",
    Cyan = "#00FFFF",
    Green = "#00FF00",
    Magenta = "#FF00FF",
    Red = "#FF0000",
    Blue = "#0000FF",
    Black = "#000000"
}